auth middleware

+++ internal/storage/storage.go
@@ -3,9 +3,10 @@ package storage
 import (
 	"context"
 	"fmt"
+	"time"
+
 	"github.com/go-code-mentor/wp-task/internal/entities"
 	"github.com/jackc/pgx/v5"
-	"time"
 )
 
 const rowsRetrieveTimeout = 10 * time.Second
@@ -24,9 +25,10 @@ type TaskSQL struct {
 	ID          uint64 `db:"id"`
 	Name        string `db:"name"`
 	Description string `db:"description"`
+	Owner       string `db:"owner"`
 }
 
-func (s *Storage) Task(ctx context.Context, id uint64) (entities.Task, error) {
+func (s *Storage) Task(ctx context.Context, id uint64, login string) (entities.Task, error) {
 	// Create context with timeout for SQL query
 	c, cancel := context.WithTimeout(ctx, rowsRetrieveTimeout)
 	defer cancel()
@@ -34,8 +36,8 @@ func (s *Storage) Task(ctx context.Context, id uint64) (entities.Task, error) {
 	var taskSQL TaskSQL
 
 	// Run SQL query
-	query := `SELECT id, name, description FROM tasks WHERE id=$1`
-	err := s.conn.QueryRow(c, query, id).Scan(&taskSQL)
+	query := `SELECT id, name, description, owner FROM tasks WHERE id=$1 AND owner=$2`
+	err := s.conn.QueryRow(c, query, id, login).Scan(&taskSQL)
 	if err != nil {
 		return entities.Task{}, fmt.Errorf("unable to get query task from storage: %w", err)
 	}
@@ -45,18 +47,19 @@ func (s *Storage) Task(ctx context.Context, id uint64) (entities.Task, error) {
 		ID:          taskSQL.ID,
 		Name:        taskSQL.Name,
 		Description: taskSQL.Description,
+		Owner:       taskSQL.Owner,
 	}
 	return task, nil
 }
 
-func (s *Storage) Tasks(ctx context.Context) ([]entities.Task, error) {
+func (s *Storage) Tasks(ctx context.Context, login string) ([]entities.Task, error) {
 	// Create context with timeout for SQL query
 	c, cancel := context.WithTimeout(ctx, rowsRetrieveTimeout)
 	defer cancel()
 
 	// Run SQL query
-	query := `SELECT id, name, description FROM tasks`
-	rows, err := s.conn.Query(c, query)
+	query := `SELECT id, name, description, owner FROM tasks WHERE owner=$1`
+	rows, err := s.conn.Query(c, query, login)
 	if err != nil {
 		return nil, fmt.Errorf("unable to get query tasks from storage: %w", err)
 	}
@@ -76,20 +79,21 @@ func (s *Storage) Tasks(ctx context.Context) ([]entities.Task, error) {
 			ID:          tasksSQL[i].ID,
 			Name:        tasksSQL[i].Name,
 			Description: tasksSQL[i].Description,
+			Owner:       tasksSQL[i].Owner,
 		}
 	}
 
 	return tasks, nil
 }
 
-func (s *Storage) TaskRemove(ctx context.Context, id uint64) error {
+func (s *Storage) TaskRemove(ctx context.Context, id uint64, login string) error {
 	// Create context with timeout for SQL query
 	c, cancel := context.WithTimeout(ctx, rowsRetrieveTimeout)
 	defer cancel()
 
 	// Run SQL query
-	query := `DELETE FROM tasks WHERE id=$1`
-	_, err := s.conn.Exec(c, query, id)
+	query := `DELETE FROM tasks WHERE id=$1 and owner=$2`
+	_, err := s.conn.Exec(c, query, id, login)
 	if err != nil {
 		return fmt.Errorf("unbale to remove task from storage: %w", err)
 	}
@@ -97,14 +101,14 @@ func (s *Storage) TaskRemove(ctx context.Context, id uint64) error {
 	return nil
 }
 
-func (s *Storage) TaskUpdate(ctx context.Context, task entities.Task) error {
+func (s *Storage) TaskUpdate(ctx context.Context, task entities.Task, login string) error {
 	// Create context with timeout for SQL query
 	c, cancel := context.WithTimeout(ctx, rowsRetrieveTimeout)
 	defer cancel()
 
 	// Run SQL query
-	query := `UPDATE tasks SET name = $1,description = $2 WHERE id = $3`
-	_, err := s.conn.Exec(c, query, task.Name, task.Description, task.ID)
+	query := `UPDATE tasks SET name = $1, description = $2 WHERE id = $3 and owner=$4`
+	_, err := s.conn.Exec(c, query, task.Name, task.Description, task.ID, login)
 	if err != nil {
 		return fmt.Errorf("unable to update task in storage: %w", err)
 	}
@@ -112,7 +116,7 @@ func (s *Storage) TaskUpdate(ctx context.Context, task entities.Task) error {
 	return nil
 }
 
-func (s *Storage) TaskAdd(ctx context.Context, task entities.Task) (uint64, error) {
+func (s *Storage) TaskAdd(ctx context.Context, task entities.Task, login string) (uint64, error) {
 	// Create context with timeout for SQL query
 	c, cancel := context.WithTimeout(ctx, rowsRetrieveTimeout)
 	defer cancel()
@@ -121,12 +125,13 @@ func (s *Storage) TaskAdd(ctx context.Context, task entities.Task) (uint64, erro
 	taskSQL := TaskSQL{
 		Name:        task.Name,
 		Description: task.Description,
+		Owner:       login,
 	}
 	var taskID int64
 
 	// Run SQL query
-	query := "INSERT INTO tasks (name, description) VALUES ($1, $2) RETURNING id"
-	err := s.conn.QueryRow(c, query, taskSQL.Name, taskSQL.Description).Scan(&taskID)
+	query := "INSERT INTO tasks (name, description, owner) VALUES ($1, $2, $3) RETURNING id"
+	err := s.conn.QueryRow(c, query, taskSQL.Name, taskSQL.Description, taskSQL.Owner).Scan(&taskID)
 	if err != nil {
 		return 0, fmt.Errorf("unable to add task to storage: %w", err)
 	}
